[refactor] AutoNumberList 전략 패턴 적용하여 구현
